api/pyproject.toml
[project]
dependencies = [
    "fastapi>=0.117.1",
    "uvicorn>=0.36.0",
    "requests>=2.32.3",
    'python-jose[cryptography]>=3.3.0',
    "python-dotenv>=1.0.0",
    "motor>=3.2.1",
    "pymongo>=4.15.3",
    "google-genai>=1.41.0",
    "jinja2>=3.1.6",
    "faker>=37.11.0",
    "aiohttp>=3.13.2",
    "psycopg2-binary>=2.9.11",
    "oracledb>=3.4.0",
]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
